Create mario.c
The purpose of this code is to fulfill the requirements set in Problem Set 1 of the 2022 session of CS50.  Note that there are two versions of this problem, a less difficult version, and a more difficult version. This code is for the more difficult version. The link below will show the prompt and requirements.

https://cs50.harvard.edu/x/2022/psets/1/mario/more/