Update credit.c
Simplified the previous version of this code. Removed non-functional lines of code and obsolete comments.

The current revision of this code passes the requirements set by the prompt. However, the card check functions (verify_amex, etc.) could be simplified. The function "countdigit" could be used to check for the card length instead of using the ranged "if" statement to check to see if the card has the correct amount of digits.