Create tideman.c
First entry for the code - the "locked()" function is not properly detecting locked edge cycles. This is because the code checks a node to see which other nodes it has locked edges with. It should be checking other nodes to see which of those has a locked edge with it. This can be thought of "looking at locked edges radiating FROM a node" rather than "looking at locked edges radiating TOWARDS a node".